Changed readGyro() to use int16_t[] instead of int[]

+++ HardwarePi.h
@@ -59,7 +59,7 @@ public:
   virtual char readChar();
   virtual double time();
   virtual void readOdometer(uint32_t &timeStamp, int32_t &wheelCount, uint32_t &dt);
-  virtual bool readGyro(int g[]);
+  virtual bool readGyro(int16_t g[]);
   virtual bool button(int pin=17);
   HardwarePiInterface(Servo& Lsteering, Servo& Lthrottle);
   virtual ~HardwarePiInterface();
